requirements.txt
cffi==1.15.0
msgpack==1.0.4
py-algorand-sdk==1.15.0
pycparser==2.21
pycryptodomex==3.15.0
PyNaCl==1.5.0
pyteal==0.13.0
python-dotenv==0.20.0
